refactor(execute/table): break execute dependency in execute/table
This breaks the dependendency on the execute package from execute/table
which then allows execute/table to be used within execute instead of the
other way around.

This opens up the possibility of using some of the table utilities
inside of the execute package.
