+++ internal/execute/groupkey/groupkey_test.go
@@ -1,4 +1,4 @@
-package execute_test
+package groupkey_test
 
 import (
 	"testing"
